+++ CHANGELOG.md
@@ -9,6 +9,7 @@
     + Added in ability to pass view_helper#fae_content_form method options that you could give to any other simple form, and connected it to form_helper#fae_input
     + Added in helper_text option for nested_table
     + Changed fae_content_form from fae_input into more generic i.input to circumvent the additional methods and clear up an issue validations.
+    + \#54548: Set on_prod to false when cloning
 
 - bugs
     + \#53380: Don't link destroyed models in change tracker